tests/wpt/tests/credential-management/credentialscontainer-prevent-silent-access.https.html

<!DOCTYPE html>
<title>Credential Management API: preventSilentAccess().</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(function() {
    return navigator.credentials.preventSilentAccess()
        .then((result) => {
            assert_equals(result, undefined);
        });
}, "navigator.credentials.preventSilentAccess() resolves with undefined.");
</script>
